# doc/api.yaml
openapi: 3.0.3 
info:
  title: WASAText
  description: |-
    This OpenAPI document describes how WASAText works.
  version: "1"
  contact:
    name: Tommaso Salvi
    email: salvi.1960045@studenti.uniroma1.it
...
servers: 
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
  schemas:
...
    Error500:
      type: object
      description: "Schema for a server error (500)."
      properties:
        error:
          type: string
          description: "The code of the error"
          pattern: "^[a-zA-Z0-9 .,!?'-]+$"
          minLength: 3
          maxLength: 1000
...
security:
  - bearerAuth: []
paths: 
...
  /messages/{Message_id}:
    parameters:
        - name: Message_id
          in: path
          required: true
          description: "The unique ID of the message to retrieve."
          schema:
            type: integer
            minimum: 1
            maximum: 1000
    get:
      tags: ["messages"]
      summary: Get a message
      description: Retrieve the details of a specific message by its unique ID.
      operationId: getMessage
      responses:
        '200':
          description: Message retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                description: Schema for the retrieved message.
                required:
                  - Message_id
                  - Sender_id
                  - Conversation_id
                  - MessageContent
                  - Timestamp
                  - Reactions
                properties:
                  Message_id:
                    type: integer
                    description: "The unique identifier of the message."
                    minimum: 1
                    maximum: 1000
                  Sender_id:
                    type: integer
                    description: "The ID of the user who sent the message."
                    minimum: 1
                    maximum: 1000
                  Conversation_id:
                    type: integer
                    description: "The ID of the user who received the message."
                    minimum: 1
                    maximum: 1000
                  MessageContent:
                    type: string
                    description: "The content of the message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
                  Timestamp:
                    type: string
                    format: date-time
                    description: "The timestamp when the message was sent."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
                  Reactions:
                    type: string
                    description: "A string containing the reactions for the message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid Message ID format or missing ID.
          content:
            application/json:
              schema:
                type: object
                description: The schema for the error of invalid or missing Id.
                properties:
                  error:
                    type: string
                    description: "Details about the error."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '403':
          description: User does not have permission to access this message.
          content:
            application/json:
              schema:
                type: object
                description: The schema for the error of invalid permission
                properties:
                  error:
                    type: string
                    description: "Permission denied."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Message not found.
          content:
            application/json:
              schema:
                type: object
                description: The schema for the error of message not found
                properties:
                  error:
                    type: string
                    description: "The message ID does not exist."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
      
    delete:
      tags: ["messages"]
      summary: Delete a message
      description: Delete a specific message
      operationId: deleteMessage
      responses:
        '200':
          description: Message deleted successfully.
          content:
            application/json:
              schema:
                type: object
                description: Confirmation that the message has been deleted.
                properties:
                  success:
                    type: boolean
                    description: "Indicates that the operation was successful."
                  message:
                    type: string
                    description: "Descriptive confirmation of the deletion."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid request. The provided `Message_id` is malformed or missing.
          content:
            application/json:
              schema:
                type: object
                description: Rejection of deleting a message.
                properties:
                  error:
                    type: string
                    description: "Error message explaining the invalid data."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                type: object
                description: Schema for permission denied
                properties:
                  error:
                    type: string
                    description: "Permission error."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Message not found.
          content:
            application/json:
              schema:
                type: object
                description: Rejection of deleting a message caused by not found
                properties:
                  error:
                    type: string
                    description: "The specified Message_id does not exist."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...
  /groups/{Group_id}/users/{User_id}:
    post:
      tags: ["groups"]
      summary: Add a user in a group
      description: Adds a user in a specific group.
      operationId: addToGroup
      parameters:
        - name: Group_id
          description: The integer that describes the Group_id.
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the group."
        - name: User_id
          description: The integer that describes the User_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the user."
      requestBody:
        description: Details of the user to add 
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for managing group users.
              required:
                - User_id
                - action
              properties:
                User_id:
                  type: integer
                  description: "ID of the user to add"
                  minimum: 1
                  maximum: 1000
      responses:
        '201':
          description: User successfully added in the group.
          content:
            application/json:
              schema:
                type: object
                description: Schema for a successful response.
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    description: "Indicates the operation was successful."
                  message:
                    type: string
                    description: "A confirmation message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid data (Group_id, User_id).
          content:
            application/json:
              schema:
                type: object
                description: Schema for invalid data.
                properties:
                  error:
                    type: string
                    description: "Details of the error."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group or user not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for error caused by not found.
                properties:
                  error:
                    type: string
                    description: "Group or user not found."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'

    delete:
      tags: ["groups"]
      summary: Remove a user by a group
      description: Removes a user in a specific group.
      operationId: leaveGroup
      parameters:
        - name: Group_id
          description: The integer that describes the Group_id.
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the group."
        - name: User_id
          description: The integer that describes the User_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            description: "The ID of the user."
      responses:
        '201':
          description: User successfully removed by the group.
          content:
            application/json:
              schema:
                type: object
                description: Schema for a successful response.
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    description: "Indicates the operation was successful."
                    example: true
                  message:
                    type: string
                    description: "A confirmation message."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '400':
          description: Invalid data (Group_id, User_id, or action).
          content:
            application/json:
              schema:
                type: object
                description: Schema for invalid data.
                properties:
                  error:
                    type: string
                    description: "Details of the error."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '404':
          description: Group or user not found.
          content:
            application/json:
              schema:
                type: object
                description: Schema for error caused by not found.
                properties:
                  error:
                    type: string
                    description: "Group or user not found."
                    pattern: "^[a-zA-Z0-9 .,!?'-]+$"
                    minLength: 3
                    maxLength: 1000
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500'
...